Extract ApiOutputFormatter and move API view types to Views/
- Move ApiTypeView, CliApiSurface, EnumValueRow, TypeParameterRow, InterfaceRow,
  BaseclassRow from ApiSurface.cs to Views/ApiViews.cs
- Move TypeShapeView + TypeViewContext from ApiCommand.cs to Views/ApiViews.cs
- Create ApiOutputFormatter in Output/ with all rendering methods:
  RenderFullApiMarkdown, RenderTypeMarkdown, WriteShapeOutput,
  RenderSignaturesOnly, plus all helper/builder methods
- ApiCommand reduced from 1105 to 387 lines (pure orchestration)
- ApiCommand no longer imports Markout
- All 253 tests pass
