+++ src/dotnet-inspect/Commands/ApiCommand.cs
@@ -1,4 +1,3 @@
-using System.Text;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 using System.Text.Json.Serialization.Metadata;
@@ -7,7 +6,7 @@ using DotnetInspector.Metadata;
 using DotnetInspector.Options;
 using DotnetInspector.Output;
 using DotnetInspector.Services;
-using Markout;
+using DotnetInspector.Views;
 
 namespace DotnetInspector.Commands;
 
@@ -321,160 +320,7 @@ public class ApiCommand
         }
         else
         {
-            Console.WriteLine(RenderFullApiMarkdown(api, options));
-        }
-    }
-
-    private static string RenderFullApiMarkdown(ApiSurface api, ApiOptions options)
-    {
-        var totalCount = api.Types.Count;
-
-        // Pre-truncate types list if --limit (before serialization so section reflects limit)
-        int? truncatedCount = null;
-        if (options.Limit.HasValue && options.Limit.Value < totalCount)
-        {
-            truncatedCount = totalCount - options.Limit.Value;
-            api.Types = api.Types.Take(options.Limit.Value).ToList();
-        }
-
-        // Single writer with section filtering
-        var writerOptions = new MarkoutWriterOptions
-        {
-            IncludeSections = options.IncludeSections,
-            ExcludeSections = options.ExcludeSections,
-            IncludeDescription = options.Verbosity != Verbosity.Quiet
-        };
-        var writer = new MarkoutWriter(writerOptions);
-
-        // Serialize title + summary fields via CLI wrapper
-        new MarkoutContext().Serialize(new CliApiSurface(api), writer);
-
-        if (totalCount == 0)
-        {
-            writer.WriteParagraph("This assembly contains no public types.");
-
-            if (api.TypeForwarders.Count > 0)
-            {
-                writer.WriteParagraph("Type forwarders could not be resolved. Target assemblies:");
-
-                var byAssembly = api.TypeForwarders
-                    .GroupBy(f => f.TargetAssembly)
-                    .OrderBy(g => g.Key)
-                    .ToList();
-
-                writer.WriteTable(
-                    new[] { "Target Assembly", "Types" },
-                    byAssembly.Select(g => new[] { g.Key, g.Count().ToString() }));
-            }
-        }
-        else
-        {
-            if (api.IsTypeForwardingAssembly)
-            {
-                writer.WriteParagraph("*This is a type-forwarding assembly. Types shown are resolved from target assemblies.*");
-            }
-
-            // Per-kind type sections (verbosity-independent)
-            RenderTypesPerKind(writer, api.Types, options);
-
-            if (truncatedCount.HasValue)
-            {
-                writer.WriteParagraph($"... *and {truncatedCount.Value} more types*");
-            }
-        }
-
-        // Source section (--docs/--samples)
-        if (options.ShowDocs || options.ShowSamples)
-            RenderAssemblySourceInfo(writer, api, options);
-
-        return writer.ToString().TrimEnd();
-    }
-
-    private static string PluralizeTypeKind(string kind) => kind switch
-    {
-        "class" => "Classes",
-        "struct" => "Structs",
-        "interface" => "Interfaces",
-        "enum" => "Enums",
-        "delegate" => "Delegates",
-        _ => char.ToUpper(kind[0]) + kind[1..] + "s"
-    };
-
-    /// <summary>
-    /// Converts CLR backtick generic names to C#-style: "Dictionary`2" → "Dictionary&lt;TKey, TValue&gt;".
-    /// Falls back to placeholder letters if TypeParameters is not populated.
-    /// </summary>
-    private static string FormatGenericTypeName(string name, List<TypeParameter>? typeParameters)
-    {
-        int backtickIndex = name.IndexOf('`');
-        if (backtickIndex < 0)
-            return name;
-
-        var baseName = name[..backtickIndex];
-        if (typeParameters is { Count: > 0 })
-            return $"{baseName}<{string.Join(", ", typeParameters.Select(tp => tp.Name))}>";
-
-        // Fallback: use arity to generate T, T1, T2, ...
-        if (int.TryParse(name[(backtickIndex + 1)..], out int arity) && arity > 0)
-        {
-            var names = arity == 1 ? "T" : string.Join(", ", Enumerable.Range(1, arity).Select(i => $"T{i}"));
-            return $"{baseName}<{names}>";
-        }
-
-        return name;
-    }
-
-    private static int GetTypeKindSortOrder(string kind) => kind switch
-    {
-        "class" => 0,
-        "struct" => 1,
-        "interface" => 2,
-        "enum" => 3,
-        "delegate" => 4,
-        _ => 5
-    };
-
-    /// <summary>
-    /// Renders per-kind type sections. Verbosity does not affect the types view;
-    /// the only variation is whether --docs adds a Description column.
-    /// </summary>
-    private static void RenderTypesPerKind(MarkoutWriter writer, List<ApiType> types, ApiOptions options)
-    {
-        bool showDocs = options.ShowDocs;
-
-        var byKind = types
-            .GroupBy(t => t.Kind)
-            .OrderBy(g => GetTypeKindSortOrder(g.Key))
-            .ToList();
-
-        foreach (var group in byKind)
-        {
-            var sectionName = PluralizeTypeKind(group.Key);
-            writer.WriteHeading(2, sectionName);
-
-            var headers = showDocs
-                ? new[] { "Type", "Members", "Description" }
-                : new[] { "Type", "Members" };
-
-            var rows = group.Select(t =>
-            {
-                var displayName = FormatGenericTypeName(t.Name, t.TypeParameters);
-                var fullName = string.IsNullOrEmpty(t.Namespace) ? displayName : $"{t.Namespace}.{displayName}";
-                var members = (t.Members?.Count ?? 0).ToString();
-
-                if (showDocs)
-                {
-                    var desc = t.Documentation?.Summary ?? "";
-                    desc = desc.Replace("\n", " ").Replace("\r", "");
-                    if (desc.Length > 80)
-                        desc = desc[..77] + "...";
-                    return new[] { fullName, members, desc };
-                }
-
-                return new[] { fullName, members };
-            });
-
-            writer.WriteTable(headers, rows);
+            Console.WriteLine(ApiOutputFormatter.RenderFullApiMarkdown(api, options));
         }
     }
 
@@ -484,7 +330,7 @@ public class ApiCommand
     {
         if (options.ShapeOutput)
         {
-            WriteShapeOutput(type, foundIn, packageName, packageVersion, options.MemberFilter);
+            ApiOutputFormatter.WriteShapeOutput(type, foundIn, packageName, packageVersion, options.MemberFilter);
         }
         else if (options.JsonOutput)
         {
@@ -492,7 +338,7 @@ public class ApiCommand
         }
         else
         {
-            Console.WriteLine(RenderTypeMarkdown(type, foundIn, packageName, packageVersion, options));
+            Console.WriteLine(ApiOutputFormatter.RenderTypeMarkdown(type, foundIn, packageName, packageVersion, options));
         }
     }
 
@@ -537,569 +383,5 @@ public class ApiCommand
             Console.WriteLine(JsonSerializer.Serialize(outputType, ApiTypeJsonContext.Default.ApiType));
     }
 
-    private static string RenderTypeMarkdown(ApiType type, string? foundIn, string? packageName, string? packageVersion, ApiOptions options)
-    {
-        // Signatures-only: plain text, no serializer
-        if (options.SignaturesOnly)
-            return RenderSignaturesOnly(type, options);
-
-        // Build the view model
-        var view = BuildApiTypeView(type, foundIn, packageName, packageVersion, options);
-
-        // Populate enum values declaratively for Normal+ enums
-        if (type.Kind == "enum" && options.Verbosity >= Verbosity.Normal)
-            PopulateEnumValues(view, type, options);
-
-        // Single writer with section filtering via MarkoutWriterOptions
-        var writerOptions = new MarkoutWriterOptions
-        {
-            IncludeSections = options.IncludeSections,
-            ExcludeSections = options.ExcludeSections,
-            IncludeDescription = options.Verbosity != Verbosity.Quiet
-        };
-        var writer = new MarkoutWriter(writerOptions);
-
-        // Serialize title + description + identity fields + enum values + type params + interfaces + baseclass
-        new MarkoutContext().Serialize(view, writer);
-
-        // Imperative rendering for member tables and source info
-        int truncatedCount = 0;
-        string truncatedNoun = "";
-
-        if (view.EnumValues == null && view.EnumValuesWithDocs == null)
-        {
-            if (options.CtorOnly && options.Verbosity >= Verbosity.Normal &&
-                type.Members?.Any(m => m.Kind == "constructor") == true)
-            {
-                // --ctor emphasis mode
-                var grouped = GroupMembersByKind(type, options.MemberFilter, options.UnsafeOnly);
-                var members = grouped
-                    .SelectMany(g => g.Value)
-                    .Where(m => m.Kind == "constructor")
-                    .ToList();
-                RenderConstructorEmphasis(writer, type, members);
-            }
-            else
-            {
-                // Per-kind member sections (all verbosity levels)
-                (truncatedCount, truncatedNoun) = RenderMembersPerKind(writer, type, options);
-            }
-        }
-
-        // Source info section (--docs/--samples)
-        if (options.ShowDocs || options.ShowSamples)
-            RenderSourceInfo(writer, type, options);
-
-        // Truncation message
-        if (truncatedCount > 0)
-            writer.WriteParagraph($"... *and {truncatedCount} more {truncatedNoun}*");
-
-        return writer.ToString().TrimEnd();
-    }
-
-    // ===== View Model Factory =====
-
-    private static ApiTypeView BuildApiTypeView(ApiType type, string? foundIn, string? packageName, string? packageVersion, ApiOptions options)
-    {
-        // Build title with package context
-        var packageInfo = packageName != null && packageVersion != null
-            ? $" ({packageName} {packageVersion})"
-            : packageName != null ? $" ({packageName})" : "";
-
-        // Build modifiers
-        var modifiers = new List<string>();
-        if (type.IsStatic) modifiers.Add("static");
-        if (type.IsAbstract && type.Kind == "class") modifiers.Add("abstract");
-        if (type.IsSealed && type.Kind == "class") modifiers.Add("sealed");
-
-        // Base type (filter out trivial bases)
-        string? baseType = null;
-        if (!string.IsNullOrEmpty(type.BaseType) && type.BaseType != "System.Object" && type.BaseType != "System.ValueType" && type.BaseType != "System.Enum")
-            baseType = type.BaseType;
-
-        // Type parameters inline (for quiet/minimal only)
-        string? typeParamsInline = null;
-        if (type.TypeParameters is { Count: > 0 } &&
-            (options.Verbosity == Verbosity.Quiet || options.Verbosity == Verbosity.Minimal))
-        {
-            var paramDescriptions = type.TypeParameters
-                .Select(tp => tp.Constraints.Count > 0
-                    ? $"{tp.DisplayName} : {tp.ConstraintsSummary}"
-                    : tp.DisplayName);
-            typeParamsInline = string.Join(", ", paramDescriptions);
-        }
-
-        // Description (from docs)
-        string? description = null;
-        if (options.ShowDocs && type.Documentation?.Summary != null)
-            description = type.Documentation.Summary;
-
-        // Samples info (only with --docs/--samples)
-        string? samplesInfo = null;
-        if ((options.ShowDocs || options.ShowSamples) && type.Documentation?.Samples?.Count > 0)
-            samplesInfo = $"{type.Documentation.Samples.Count} available";
-
-        // Type parameters table (Normal+)
-        List<TypeParameterRow>? typeParameterRows = null;
-        if (type.TypeParameters is { Count: > 0 } && options.Verbosity >= Verbosity.Normal)
-        {
-            typeParameterRows = type.TypeParameters
-                .Select(tp => new TypeParameterRow { Parameter = tp.DisplayName, Constraints = tp.ConstraintsSummary ?? "" })
-                .ToList();
-        }
-
-        // Interfaces (Detailed+)
-        List<InterfaceRow>? interfaceRows = null;
-        if (type.Interfaces is { Count: > 0 } && options.Verbosity >= Verbosity.Detailed)
-        {
-            interfaceRows = type.Interfaces.Order()
-                .Select(i => new InterfaceRow { Interface = i })
-                .ToList();
-        }
-
-        // Baseclass (Detailed+, filtered for trivial bases)
-        List<BaseclassRow>? baseclassRows = null;
-        if (baseType != null && options.Verbosity >= Verbosity.Detailed)
-        {
-            baseclassRows = [new BaseclassRow { Type = baseType }];
-        }
-
-        return new ApiTypeView
-        {
-            Title = $"{type.FullName}{packageInfo}",
-            Description = description,
-            Kind = type.Kind,
-            Modifiers = modifiers.Count > 0 ? string.Join(", ", modifiers) : null,
-            BaseType = baseType,
-            TypeParametersInline = typeParamsInline,
-            Assembly = foundIn,
-            Package = packageName,
-            Version = packageVersion,
-            SamplesInfo = samplesInfo,
-            TypeParameterRows = typeParameterRows,
-            InterfaceRows = interfaceRows,
-            BaseclassRows = baseclassRows
-        };
-    }
-
-    // ===== Member Rendering =====
-
-    /// <summary>
-    /// Populates enum value rows on the view model for declarative serialization.
-    /// </summary>
-    private static void PopulateEnumValues(ApiTypeView view, ApiType type, ApiOptions options)
-    {
-        var enumMembers = (type.Members ?? [])
-            .Where(m => m.Kind == "field" && m.EnumValue.HasValue && !IsCompilerGenerated(m.Name))
-            .OrderBy(m => m.EnumValue)
-            .ToList();
-
-        if (enumMembers.Count == 0)
-            return;
-
-        if (options.Limit.HasValue && options.Limit.Value < enumMembers.Count)
-            enumMembers = enumMembers.Take(options.Limit.Value).ToList();
-
-        bool hasAnyDocs = options.ShowDocs && enumMembers.Any(m => m.Documentation?.Summary != null);
-
-        var rows = enumMembers.Select(m => new EnumValueRow
-        {
-            Name = m.Name,
-            Value = m.EnumValue.ToString()!,
-            Description = hasAnyDocs ? (m.Documentation?.Summary ?? "") : null
-        }).ToList();
-
-        if (hasAnyDocs)
-            view.EnumValuesWithDocs = rows;
-        else
-            view.EnumValues = rows;
-    }
-
-    /// <summary>
-    /// Writes per-kind member section tables. Returns (truncated, noun) for truncation message.
-    /// </summary>
-    private static (int truncated, string noun) RenderMembersPerKind(
-        MarkoutWriter writer, ApiType type, ApiOptions options)
-    {
-        var grouped = GroupMembersByKind(type, options.MemberFilter, options.UnsafeOnly);
-        if (grouped.Count == 0) return (0, "");
-
-        // Flatten sorted for --limit application
-        var allMembers = grouped
-            .SelectMany(g => g.Value)
-            .OrderBy(m => GetMemberSortOrder(m.Kind))
-            .ThenBy(m => m.Name)
-            .ToList();
-
-        int truncated = 0;
-        if (options.Limit.HasValue && options.Limit.Value < allMembers.Count)
-        {
-            truncated = allMembers.Count - options.Limit.Value;
-            allMembers = allMembers.Take(options.Limit.Value).ToList();
-        }
-
-        // Re-group after truncation
-        var kindGroups = allMembers
-            .GroupBy(m => m.Kind)
-            .OrderBy(g => GetMemberSortOrder(g.Key))
-            .ToList();
-
-        bool hasDocs = options.ShowDocs && allMembers.Any(m => m.Documentation?.Summary != null);
-        var formatter = MemberTableFormatter.Create(options.Verbosity);
-
-        foreach (var group in kindGroups)
-        {
-            var kind = group.Key;
-            var sectionName = PluralizeKind(kind);
-            var members = group.ToList();
-
-            writer.WriteHeading(2, sectionName);
-            writer.WriteTable(
-                formatter.GetHeaders(kind, members, hasDocs),
-                formatter.FormatRows(kind, members, hasDocs));
-        }
-
-        return (truncated, "members");
-    }
-
-    /// <summary>
-    /// Renders source information as a separate section (--docs/--samples).
-    /// </summary>
-    private static void RenderSourceInfo(MarkoutWriter writer, ApiType type, ApiOptions options)
-    {
-        var resolution = type.SourceResolution;
-        string? primaryUrl = type.GitHubBrowseUrl != null
-            ? (options.BrowsableUrls ? GitHubUrlResolver.ConvertRawToBlobUrl(type.GitHubBrowseUrl) : type.GitHubBrowseUrl)
-            : null;
-
-        if (resolution == null && primaryUrl == null)
-            return;
-
-        bool isPartial = type.IsPartialType;
-        int fileCount = 1 + (type.AdditionalSourceFiles?.Count ?? 0);
-
-        var rows = new List<string[]>();
-        if (!string.IsNullOrEmpty(resolution))
-            rows.Add(new[] { "Resolution", resolution });
-        rows.Add(new[] { "Partial Type", isPartial ? "true" : "false" });
-        rows.Add(new[] { "Files", fileCount.ToString() });
-
-        // Primary source file
-        if (!string.IsNullOrEmpty(primaryUrl))
-        {
-            var fileName = Path.GetFileName(type.SourceFilePath) ?? Path.GetFileName(primaryUrl);
-            rows.Add(new[] { fileName, primaryUrl });
-        }
-
-        // Additional source files for partial types
-        if (type.AdditionalSourceFiles != null)
-        {
-            foreach (var file in type.AdditionalSourceFiles)
-            {
-                var url = file.GitHubBrowseUrl != null
-                    ? (options.BrowsableUrls ? GitHubUrlResolver.ConvertRawToBlobUrl(file.GitHubBrowseUrl) : file.GitHubBrowseUrl)
-                    : file.SourceUrl ?? "";
-                var fileName = Path.GetFileName(file.FilePath) ?? Path.GetFileName(url);
-                rows.Add(new[] { fileName, url });
-            }
-        }
-
-        writer.WriteHeading(2, "Source");
-        writer.WriteTable(new[] { "Property", "Value" }, rows);
-    }
-
-    /// <summary>
-    /// Renders a simple Source section for the full-assembly view (Resolution + URL).
-    /// </summary>
-    private static void RenderAssemblySourceInfo(MarkoutWriter writer, ApiSurface api, ApiOptions options)
-    {
-        var resolution = api.Types.FirstOrDefault(t => t.SourceResolution != null)?.SourceResolution;
-        if (resolution == null && string.IsNullOrEmpty(api.RepositoryUrl))
-            return;
-
-        var rows = new List<string[]>();
-        if (!string.IsNullOrEmpty(resolution))
-            rows.Add(new[] { "Resolution", resolution });
-        if (!string.IsNullOrEmpty(api.RepositoryUrl))
-            rows.Add(new[] { "Repository", api.RepositoryUrl });
-
-        writer.WriteHeading(2, "Source");
-        writer.WriteTable(new[] { "Property", "Value" }, rows);
-    }
-
-
-    // ===== Signatures-Only Mode =====
-
-    private static string RenderSignaturesOnly(ApiType type, ApiOptions options)
-    {
-        var members = type.Members?
-            .Where(m => !IsCompilerGenerated(m.Name))
-            .OrderBy(m => GetMemberSortOrder(m.Kind))
-            .ThenBy(m => m.Name)
-            .ToList() ?? [];
-
-        if (options.MemberFilter?.Count > 0)
-            members = members.Where(m => TypeMatcher.MatchesMemberFilter(m.Name, options.MemberFilter)).ToList();
-
-        if (options.UnsafeOnly)
-            members = members.Where(m => m.IsUnsafe).ToList();
-
-        var sb = new StringBuilder();
-        var displayMembers = members.AsEnumerable();
-        if (options.Limit.HasValue && options.Limit.Value < members.Count)
-            displayMembers = displayMembers.Take(options.Limit.Value);
-
-        if (type.Kind == "enum")
-        {
-            var enumMembers = members
-                .Where(m => m.Kind == "field" && m.EnumValue.HasValue)
-                .OrderBy(m => m.EnumValue);
-            foreach (var member in options.Limit.HasValue && options.Limit.Value < members.Count
-                ? enumMembers.Take(options.Limit.Value)
-                : enumMembers)
-            {
-                sb.AppendLine($"{member.Name} = {member.EnumValue}");
-            }
-        }
-        else
-        {
-            foreach (var member in displayMembers)
-            {
-                sb.AppendLine(member.Signature ?? member.ReturnType ?? "");
-            }
-        }
-        return sb.ToString().TrimEnd();
-    }
-
-    // ===== Rendering Helpers =====
-
-    private static void RenderConstructorEmphasis(MarkoutWriter writer, ApiType type, List<ApiMember> constructors)
-    {
-        writer.WriteHeading(2, $"Constructors ({constructors.Count} overload{(constructors.Count != 1 ? "s" : "")})");
-
-        var sorted = constructors
-            .OrderBy(c => SignatureParser.CountParameters(c.Signature))
-            .ToList();
-
-        for (int i = 0; i < sorted.Count; i++)
-        {
-            var ctor = sorted[i];
-            var paramCount = SignatureParser.CountParameters(ctor.Signature);
-            var paramInfo = SignatureParser.ExtractParameterInfo(ctor.Signature);
-
-            writer.WriteHeading(3, $"Overload {i + 1}: {paramCount} parameter{(paramCount != 1 ? "s" : "")}");
-
-            writer.WriteCodeBlockStart("csharp");
-            writer.WriteParagraph($"new {type.Name}{SignatureParser.FormatConstructorCall(ctor.Signature)}");
-            writer.WriteCodeBlockEnd();
-
-            if (paramInfo.Count > 0)
-            {
-                var headers = new[] { "Parameter", "Type", "Notes" };
-                var rows = paramInfo.Select(p => new[] { p.name, $"`{p.type}`", p.hasDefault ? "optional" : "required" });
-                writer.WriteTable(headers, rows);
-            }
-        }
-    }
-
-    // ===== Helper Methods (delegated to ApiCommand where static, or copied for private ones) =====
-
-    private static Dictionary<string, List<ApiMember>> GroupMembersByKind(ApiType type, HashSet<string>? memberFilter = null, bool unsafeOnly = false)
-    {
-        var members = type.Members?
-            .Where(m => !IsCompilerGenerated(m.Name))
-            .ToList() ?? [];
-
-        if (memberFilter?.Count > 0)
-            members = members.Where(m => TypeMatcher.MatchesMemberFilter(m.Name, memberFilter)).ToList();
-
-        if (unsafeOnly)
-            members = members.Where(m => m.IsUnsafe).ToList();
-
-        return members
-            .GroupBy(m => m.Kind)
-            .OrderBy(g => GetMemberSortOrder(g.Key))
-            .ToDictionary(g => g.Key, g => g.ToList());
-    }
-
-    private static string PluralizeKind(string kind) => kind switch
-    {
-        "property" => "Properties",
-        "method" => "Methods",
-        "field" => "Fields",
-        "event" => "Events",
-        "constructor" => "Constructors",
-        _ => char.ToUpper(kind[0]) + kind[1..] + "s"
-    };
-
-    private static bool IsCompilerGenerated(string name) => MemberFilters.IsCompilerGenerated(name);
-
-    private static readonly string[] MemberKinds = ["constructor", "field", "property", "method", "event"];
-
-    private static int GetMemberSortOrder(string kind)
-    {
-        var index = Array.IndexOf(MemberKinds, kind);
-        return index >= 0 ? index : MemberKinds.Length;
-    }
-
-    // ===== Shape Output (--shape) =====
-
-    private static void WriteShapeOutput(ApiType type, string? foundIn, string? packageName, string? packageVersion, HashSet<string>? memberFilter)
-    {
-        var view = BuildShapeView(type, foundIn, packageName, packageVersion, memberFilter);
-        MarkoutSerializer.Serialize(view, Console.Out, TypeViewContext.Default);
-    }
-
-    private static TypeShapeView BuildShapeView(ApiType type, string? foundIn, string? packageName, string? packageVersion, HashSet<string>? memberFilter)
-    {
-        var nodes = new List<TreeNode>();
-
-        // Inheritance (always show)
-        if (!string.IsNullOrEmpty(type.BaseType) && type.BaseType != "Object")
-        {
-            nodes.Add(new TreeNode("Inherits", new[] { type.BaseType }));
-        }
-
-        // Interfaces (always show)
-        if (type.Interfaces is { Count: > 0 })
-        {
-            nodes.Add(new TreeNode("Implements", type.Interfaces));
-        }
-
-        // Type parameters with constraints (always show)
-        if (type.TypeParameters is { Count: > 0 })
-        {
-            var typeParamDescriptions = type.TypeParameters
-                .Select(tp => tp.Constraints.Count > 0
-                    ? $"{tp.DisplayName} : {tp.ConstraintsSummary}"
-                    : tp.DisplayName)
-                .ToList();
-            nodes.Add(new TreeNode("Type Parameters", typeParamDescriptions));
-        }
-
-        // Group members by kind
-        if (type.Members is { Count: > 0 })
-        {
-            var members = type.Members.Where(m => !IsCompilerGenerated(m.Name));
-
-            if (memberFilter?.Count > 0)
-            {
-                members = members.Where(m => TypeMatcher.MatchesMemberFilter(m.Name, memberFilter));
-            }
-
-            var membersByKind = members
-                .GroupBy(m => m.Kind)
-                .OrderBy(g => GetTreeKindOrder(g.Key));
-
-            foreach (var group in membersByKind)
-            {
-                var kindLabel = GetTreeKindLabel(group.Key, group.Count());
-                var memberSignatures = group
-                    .OrderBy(m => m.Name)
-                    .Select(m => m.Signature ?? m.Name)
-                    .ToList();
-
-                nodes.Add(new TreeNode(kindLabel, memberSignatures));
-            }
-        }
-
-        var modifiers = new List<string>();
-        if (type.IsStatic) modifiers.Add("static");
-        if (type.IsAbstract && type.Kind == "class") modifiers.Add("abstract");
-        if (type.IsSealed && type.Kind == "class") modifiers.Add("sealed");
-
-        var packageInfo = packageName != null && packageVersion != null
-            ? $" ({packageName} {packageVersion})"
-            : packageName != null ? $" ({packageName})" : "";
-
-        return new TypeShapeView
-        {
-            FullName = $"{type.FullName}{packageInfo}",
-            Kind = type.Kind,
-            Modifiers = modifiers.Count > 0 ? string.Join(", ", modifiers) : null,
-            Assembly = foundIn,
-            Package = packageName,
-            Version = packageVersion,
-            Members = nodes
-        };
-    }
-
-    private static int GetTreeKindOrder(string kind) => kind switch
-    {
-        "constructor" => 0,
-        "property" => 1,
-        "method" => 2,
-        "event" => 3,
-        "field" => 4,
-        _ => 5
-    };
-
-    private static string GetTreeKindLabel(string kind, int count)
-    {
-        var plural = kind switch
-        {
-            "property" => "Properties",
-            "method" => "Methods",
-            "constructor" => "Constructors",
-            "event" => "Events",
-            "field" => "Fields",
-            _ => kind + "s"
-        };
-        return $"{plural} ({count})";
-    }
 }
 
-/// <summary>
-/// View model for type shape output (--shape).
-/// </summary>
-[MarkoutSerializable(TitleProperty = nameof(FullName))]
-public class TypeShapeView
-{
-    [MarkoutIgnore]
-    public string FullName { get; set; } = "";
-
-    [MarkoutIgnore]
-    public string Kind { get; set; } = "";
-
-    [MarkoutIgnore]
-    public string? Modifiers { get; set; }
-
-    [MarkoutIgnore]
-    public string? Assembly { get; set; }
-
-    [MarkoutIgnore]
-    public string? Package { get; set; }
-
-    [MarkoutIgnore]
-    public string? Version { get; set; }
-
-    /// <summary>
-    /// Identity fields (rendered as key-value fields under the title).
-    /// </summary>
-    [JsonIgnore]
-    [MarkoutIgnoreInTable]
-    public List<MarkoutField> Identity => GetIdentityFields();
-
-    [MarkoutIgnoreInTable]
-    public List<TreeNode> Members { get; set; } = [];
-
-    private List<MarkoutField> GetIdentityFields()
-    {
-        var fields = new List<MarkoutField>();
-        fields.Add(new("Kind", Kind));
-        if (!string.IsNullOrEmpty(Modifiers))
-            fields.Add(new("Modifiers", Modifiers));
-        if (!string.IsNullOrEmpty(Assembly))
-            fields.Add(new("Assembly", Assembly));
-        if (!string.IsNullOrEmpty(Package))
-            fields.Add(new("Package", Package));
-        if (!string.IsNullOrEmpty(Version))
-            fields.Add(new("Version", Version));
-        return fields;
-    }
-}
-
-[MarkoutContext(typeof(TypeShapeView))]
-public partial class TypeViewContext : MarkoutSerializerContext
-{
-}
